friend mood map working, cleaning up code
