+++ moodBookProject/app/src/main/java/com/example/moodbook/ui/myFriendMoodMap/MyFriendMoodMapFragment.java
@@ -17,7 +17,6 @@ import com.example.moodbook.MoodMapFragment;
 import com.example.moodbook.Mood;
 import com.example.moodbook.R;
 import com.example.moodbook.ui.friendMood.FriendMood;
-import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.MapView;
 import com.google.android.gms.maps.OnMapReadyCallback;
@@ -83,9 +82,6 @@ public class MyFriendMoodMapFragment extends MoodMapFragment implements OnMapRea
         // initialize map
         moodMap = googleMap;
 
-        // update list of markers
-        updateList(db);
-
         mapView.setContentDescription("MAP READY");
 
         moodMap.setOnMarkerClickListener(new GoogleMap.OnMarkerClickListener() {
@@ -131,39 +127,9 @@ public class MyFriendMoodMapFragment extends MoodMapFragment implements OnMapRea
         super.onLowMemory();
         mapView.onLowMemory();
     }
-    @Override
-    public void updateList(final FirebaseFirestore db) {
-    /*
-        try {
-            db.collection("USERS")
-                    .document(userID)
-                    .collection("FRIENDS")
-                    .get()
-                    .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                        @Override
-                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                            if(task.isSuccessful()){
-                                for(int i = 0; i < task.getResult().size(); i++){
-                                    DocumentSnapshot doc = task.getResult().getDocuments().get(i);
-
-                                    if (doc.exists() && !(doc.getId().equals("null"))){
-                                        Log.i("DOCID", "id: " + doc.getId() + "  |  " + doc.get("uid"));
-
-
-                                    }
-                                }
-                            }
-                        }
-                    });
-        } catch (Exception e){
-            e.printStackTrace();
-        }
-    */
-    }
 
 
     private void drawMood(Mood mood, int i ){
-        //todo: check if location exists in mood
         if (mood.getLocation() != null){
             // get image resource for the mood marker
             int emotionResource = mood.getEmotionImageResource();
